// ng-jz-paint/src/app/interfaces/room-interface.ts
export interface RoomInterface {

    roomName: string;
    wallLength: string;
    wallWidth: string;
    wallHeight: string;
    wallArea: string;
    ceilingLength: string;
    ceilingWidth: string;
    ceilingArea: string;
    trimLength: string;
    trimWidth: string;
    trimArea: string;
    numDoorFaces: string;
    numDoorFrames: string;
    numWindowFrames: string;
    numAddons: string;

    wallPrice: string;
    ceilingPrice: string;
    trimPrice: string;
    doorFacePrice: string;
    doorFramePrice: string;
    windowFramePrice: string;
    addonTotalPrice: string;

    totalRoomPrice: string;

}